POL-152: calendar file move to component
